requirements.txt
python-chrisclient

# dev
pytest==7.2.0
pytest-cov==3.0.0